Move theme button and remove left drawer
